ItemAssign 切り出し
